0.005 complete rewrite
Network class separated from the Client and Server to allow for custom game loops to work directly with the Network. Messenger is included for the Client and Server but not needed for the Network itself, which provides all of the core functionality. Clock is being used to nicely handle logging but that hasn't been fully fleshed out/implemented so not sure if that's how it's going to be.